requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==4.2.23
django-jazzmin==3.0.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
pillow==11.3.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
djangorestframework-simplejwt>=5.3.0
django-redis==5.4.0
redis==5.0.3
celery==5.3.6
python-dotenv==1.0.0
psycopg==3.1.18


